More Time Work
Configured the main clock for 1MHz.  Set up Timer1_A0 to interrupt at
50ms intervals and count the system_time at .1 second intervals.  The
idea is that we don't really need accurate time, so .1s resolution is
way more than good.  Also re-worked the timer functions, because the
previous idea was flawed.  When the system_time variable overflows, the
timer datastructure needs to know about the overflow so it can correctly
determine if the timer is expired.